src/Alarm/Alarm/requirements.txt
django
djangorestframework
requests
bs4
django-apscheduler
gunicorn